Create homes for domain code and drivers
