+++ desktop/features/support/hooks.js
@@ -4,10 +4,16 @@ const path = require('path');
 const Application = require("spectron").Application
 const electronPath = require("electron")
 const appPath = path.join(__dirname, '../../')
-const {Before, After, setDefaultTimeout} = require('@cucumber/cucumber')
+const { Before, After, setDefaultTimeout } = require('@cucumber/cucumber')
+const { Setup, UI } = require("../support/drivers");
 
 // uncomment to debug without steps timing out
 // setDefaultTimeout(-1);
+
+Before(function() {
+  this.setup = new Setup();
+  this.ui = new UI();
+});
 
 Before({tags: "@ui"}, function() {
   this.app = new Application({
